sync with upstream

+++ src/org/thoughtcrime/securesms/jobs/MmsDownloadJob.java
@@ -2,62 +2,74 @@ package org.thoughtcrime.securesms.jobs;
 
 import android.content.Context;
 import android.net.Uri;
-import android.util.Log;
-import android.util.Pair;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+
+import org.thoughtcrime.securesms.jobmanager.SafeData;
+import org.thoughtcrime.securesms.logging.Log;
+
+import com.google.android.mms.pdu_alt.CharacterSets;
+import com.google.android.mms.pdu_alt.EncodedStringValue;
+import com.google.android.mms.pdu_alt.PduBody;
+import com.google.android.mms.pdu_alt.PduPart;
+import com.google.android.mms.pdu_alt.RetrieveConf;
 
 import org.thoughtcrime.securesms.attachments.Attachment;
 import org.thoughtcrime.securesms.attachments.UriAttachment;
 import org.thoughtcrime.securesms.crypto.MasterSecret;
-import org.thoughtcrime.securesms.crypto.MasterSecretUnion;
-import org.thoughtcrime.securesms.database.DatabaseFactory;
-import org.thoughtcrime.securesms.database.MmsDatabase;
+import org.thoughtcrime.securesms.database.Address;
 import org.thoughtcrime.securesms.database.AttachmentDatabase;
-import org.thoughtcrime.securesms.jobs.requirements.MasterSecretRequirement;
+import org.thoughtcrime.securesms.database.DatabaseFactory;
+import org.thoughtcrime.securesms.database.MessagingDatabase.InsertResult;
+import org.thoughtcrime.securesms.database.MmsDatabase;
+import org.thoughtcrime.securesms.jobmanager.JobParameters;
 import org.thoughtcrime.securesms.mms.ApnUnavailableException;
 import org.thoughtcrime.securesms.mms.CompatMmsConnection;
 import org.thoughtcrime.securesms.mms.IncomingMediaMessage;
+import org.thoughtcrime.securesms.mms.MmsException;
 import org.thoughtcrime.securesms.mms.MmsRadioException;
 import org.thoughtcrime.securesms.mms.PartParser;
 import org.thoughtcrime.securesms.notifications.MessageNotifier;
-import org.thoughtcrime.securesms.providers.SingleUseBlobProvider;
+import org.thoughtcrime.securesms.providers.MemoryBlobProvider;
 import org.thoughtcrime.securesms.service.KeyCachingService;
+import org.thoughtcrime.securesms.util.TextSecurePreferences;
 import org.thoughtcrime.securesms.util.Util;
-import org.whispersystems.jobqueue.JobParameters;
-import org.whispersystems.jobqueue.requirements.NetworkRequirement;
-import org.whispersystems.libaxolotl.DuplicateMessageException;
-import org.whispersystems.libaxolotl.InvalidMessageException;
-import org.whispersystems.libaxolotl.LegacyMessageException;
-import org.whispersystems.libaxolotl.NoSessionException;
-import org.whispersystems.libaxolotl.util.guava.Optional;
+import org.whispersystems.libsignal.DuplicateMessageException;
+import org.whispersystems.libsignal.InvalidMessageException;
+import org.whispersystems.libsignal.LegacyMessageException;
+import org.whispersystems.libsignal.NoSessionException;
+import org.whispersystems.libsignal.util.guava.Optional;
 
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
+import java.util.Set;
 
-import ws.com.google.android.mms.ContentType;
-import ws.com.google.android.mms.MmsException;
-import ws.com.google.android.mms.pdu.EncodedStringValue;
-import ws.com.google.android.mms.pdu.NotificationInd;
-import ws.com.google.android.mms.pdu.PduBody;
-import ws.com.google.android.mms.pdu.PduPart;
-import ws.com.google.android.mms.pdu.RetrieveConf;
+import androidx.work.Data;
 
-public class MmsDownloadJob extends MasterSecretJob {
+public class MmsDownloadJob extends ContextJob {
+
+  private static final long serialVersionUID = 1L;
 
   private static final String TAG = MmsDownloadJob.class.getSimpleName();
 
-  private final long    messageId;
-  private final long    threadId;
-  private final boolean automatic;
+  private static final String KEY_MESSAGE_ID = "message_id";
+  private static final String KEY_THREAD_ID  = "thread_id";
+  private static final String KEY_AUTOMATIC  = "automatic";
+
+  private long    messageId;
+  private long    threadId;
+  private boolean automatic;
+
+  public MmsDownloadJob() {
+    super(null, null);
+  }
 
   public MmsDownloadJob(Context context, long messageId, long threadId, boolean automatic) {
     super(context, JobParameters.newBuilder()
-                                .withPersistence()
-                                .withRequirement(new MasterSecretRequirement(context))
-                                .withRequirement(new NetworkRequirement(context))
                                 .withGroupId("mms-operation")
-                                .withWakeLock(true, 30, TimeUnit.SECONDS)
                                 .create());
 
     this.messageId = messageId;
@@ -66,17 +78,32 @@ public class MmsDownloadJob extends MasterSecretJob {
   }
 
   @Override
+  protected void initialize(@NonNull SafeData data) {
+    messageId = data.getLong(KEY_MESSAGE_ID);
+    threadId  = data.getLong(KEY_THREAD_ID);
+    automatic = data.getBoolean(KEY_AUTOMATIC);
+  }
+
+  @Override
+  protected @NonNull Data serialize(@NonNull Data.Builder dataBuilder) {
+    return dataBuilder.putLong(KEY_MESSAGE_ID, messageId)
+                      .putLong(KEY_THREAD_ID, threadId)
+                      .putBoolean(KEY_AUTOMATIC, automatic)
+                      .build();
+  }
+
+  @Override
   public void onAdded() {
-    if (automatic && KeyCachingService.getMasterSecret(context) == null) {
+    if (automatic && KeyCachingService.isLocked(context)) {
       DatabaseFactory.getMmsDatabase(context).markIncomingNotificationReceived(threadId);
-      MessageNotifier.updateNotification(context, null);
+      MessageNotifier.updateNotification(context);
     }
   }
 
   @Override
-  public void onRun(MasterSecret masterSecret) {
-    MmsDatabase               database     = DatabaseFactory.getMmsDatabase(context);
-    Optional<NotificationInd> notification = database.getNotification(messageId);
+  public void onRun() {
+    MmsDatabase                               database     = DatabaseFactory.getMmsDatabase(context);
+    Optional<MmsDatabase.MmsNotificationInfo> notification = database.getNotification(messageId);
 
     if (!notification.isPresent()) {
       Log.w(TAG, "No notification for ID: " + messageId);
@@ -88,30 +115,46 @@ public class MmsDownloadJob extends MasterSecretJob {
         throw new MmsException("Notification content location was null.");
       }
 
+      if (!TextSecurePreferences.isPushRegistered(context)) {
+        throw new MmsException("Not registered");
+      }
+
       database.markDownloadState(messageId, MmsDatabase.Status.DOWNLOAD_CONNECTING);
 
-      String contentLocation = new String(notification.get().getContentLocation());
-      byte[] transactionId   = notification.get().getTransactionId();
+      String contentLocation = notification.get().getContentLocation();
+      byte[] transactionId   = new byte[0];
 
-      Log.w(TAG, "Downloading mms at " + Uri.parse(contentLocation).getHost());
+      try {
+        if (notification.get().getTransactionId() != null) {
+          transactionId = notification.get().getTransactionId().getBytes(CharacterSets.MIMENAME_ISO_8859_1);
+        } else {
+          Log.w(TAG, "No transaction ID!");
+        }
+      } catch (UnsupportedEncodingException e) {
+        Log.w(TAG, e);
+      }
 
-      RetrieveConf retrieveConf = new CompatMmsConnection(context).retrieve(contentLocation, transactionId);
+      Log.i(TAG, "Downloading mms at " + Uri.parse(contentLocation).getHost() + ", subscription ID: " + notification.get().getSubscriptionId());
+
+      RetrieveConf retrieveConf = new CompatMmsConnection(context).retrieve(contentLocation, transactionId, notification.get().getSubscriptionId());
+
       if (retrieveConf == null) {
         throw new MmsException("RetrieveConf was null");
       }
-      storeRetrievedMms(masterSecret, contentLocation, messageId, threadId, retrieveConf);
+
+      storeRetrievedMms(contentLocation, messageId, threadId, retrieveConf, notification.get().getSubscriptionId(), notification.get().getFrom());
     } catch (ApnUnavailableException e) {
       Log.w(TAG, e);
-      handleDownloadError(masterSecret, messageId, threadId, MmsDatabase.Status.DOWNLOAD_APN_UNAVAILABLE,
+      handleDownloadError(messageId, threadId, MmsDatabase.Status.DOWNLOAD_APN_UNAVAILABLE,
                           automatic);
     } catch (MmsException e) {
       Log.w(TAG, e);
-      handleDownloadError(masterSecret, messageId, threadId,
+      handleDownloadError(messageId, threadId,
                           MmsDatabase.Status.DOWNLOAD_HARD_FAILURE,
                           automatic);
     } catch (MmsRadioException | IOException e) {
       Log.w(TAG, e);
-      handleDownloadError(masterSecret, messageId, threadId,
+      handleDownloadError(messageId, threadId,
                           MmsDatabase.Status.DOWNLOAD_SOFT_FAILURE,
                           automatic);
     } catch (DuplicateMessageException e) {
@@ -136,43 +179,52 @@ public class MmsDownloadJob extends MasterSecretJob {
 
     if (automatic) {
       database.markIncomingNotificationReceived(threadId);
-      MessageNotifier.updateNotification(context, null, threadId);
+      MessageNotifier.updateNotification(context, threadId);
     }
   }
 
   @Override
-  public boolean onShouldRetryThrowable(Exception exception) {
+  public boolean onShouldRetry(Exception exception) {
     return false;
   }
 
-  private void storeRetrievedMms(MasterSecret masterSecret, String contentLocation,
-                                 long messageId, long threadId, RetrieveConf retrieved)
+  private void storeRetrievedMms(String contentLocation,
+                                 long messageId, long threadId, RetrieveConf retrieved,
+                                 int subscriptionId, @Nullable Address notificationFrom)
       throws MmsException, NoSessionException, DuplicateMessageException, InvalidMessageException,
              LegacyMessageException
   {
     MmsDatabase           database    = DatabaseFactory.getMmsDatabase(context);
-    SingleUseBlobProvider provider    = SingleUseBlobProvider.getInstance();
-    String                from        = null;
-    List<String>          to          = new LinkedList<>();
-    List<String>          cc          = new LinkedList<>();
+    MemoryBlobProvider    provider    = MemoryBlobProvider.getInstance();
+    Optional<Address>     group       = Optional.absent();
+    Set<Address>          members     = new HashSet<>();
     String                body        = null;
     List<Attachment>      attachments = new LinkedList<>();
 
+    Address               from;
+
     if (retrieved.getFrom() != null) {
-      from = Util.toIsoString(retrieved.getFrom().getTextString());
+      from = Address.fromExternal(context, Util.toIsoString(retrieved.getFrom().getTextString()));
+    } else if (notificationFrom != null) {
+      from = notificationFrom;
+    } else {
+      from = Address.UNKNOWN;
     }
 
     if (retrieved.getTo() != null) {
       for (EncodedStringValue toValue : retrieved.getTo()) {
-        to.add(Util.toIsoString(toValue.getTextString()));
+        members.add(Address.fromExternal(context, Util.toIsoString(toValue.getTextString())));
       }
     }
 
     if (retrieved.getCc() != null) {
       for (EncodedStringValue ccValue : retrieved.getCc()) {
-        cc.add(Util.toIsoString(ccValue.getTextString()));
+        members.add(Address.fromExternal(context, Util.toIsoString(ccValue.getTextString())));
       }
     }
+
+    members.add(from);
+    members.add(Address.fromExternal(context, TextSecurePreferences.getLocalNumber(context)));
 
     if (retrieved.getBody() != null) {
       body = PartParser.getMessageText(retrieved.getBody());
@@ -182,26 +234,32 @@ public class MmsDownloadJob extends MasterSecretJob {
         PduPart part = media.getPart(i);
 
         if (part.getData() != null) {
-          Uri uri = provider.createUri(part.getData());
+          Uri    uri  = provider.createSingleUseUri(part.getData());
+          String name = null;
+
+          if (part.getName() != null) name = Util.toIsoString(part.getName());
+
           attachments.add(new UriAttachment(uri, Util.toIsoString(part.getContentType()),
                                             AttachmentDatabase.TRANSFER_PROGRESS_DONE,
-                                            part.getData().length));
+                                            part.getData().length, name, false, false, null));
         }
       }
     }
 
+    if (members.size() > 2) {
+      group = Optional.of(Address.fromSerialized(DatabaseFactory.getGroupDatabase(context).getOrCreateGroupForMembers(new LinkedList<>(members), true)));
+    }
 
+    IncomingMediaMessage   message      = new IncomingMediaMessage(from, group, body, retrieved.getDate() * 1000L, attachments, subscriptionId, 0, false, false);
+    Optional<InsertResult> insertResult = database.insertMessageInbox(message, contentLocation, threadId);
 
-    IncomingMediaMessage message  = new IncomingMediaMessage(from, to, cc, body, retrieved.getDate() * 1000L, attachments);
-
-    Pair<Long, Long> messageAndThreadId  = database.insertMessageInbox(new MasterSecretUnion(masterSecret),
-                                                                       message, contentLocation, threadId);
-    database.delete(messageId);
-    MessageNotifier.updateNotification(context, masterSecret, messageAndThreadId.second);
+    if (insertResult.isPresent()) {
+      database.delete(messageId);
+      MessageNotifier.updateNotification(context, insertResult.get().getThreadId());
+    }
   }
 
-  private void handleDownloadError(MasterSecret masterSecret, long messageId, long threadId,
-                                   int downloadStatus, boolean automatic)
+  private void handleDownloadError(long messageId, long threadId, int downloadStatus, boolean automatic)
   {
     MmsDatabase db = DatabaseFactory.getMmsDatabase(context);
 
@@ -209,7 +267,7 @@ public class MmsDownloadJob extends MasterSecretJob {
 
     if (automatic) {
       db.markIncomingNotificationReceived(threadId);
-      MessageNotifier.updateNotification(context, masterSecret, threadId);
+      MessageNotifier.updateNotification(context, threadId);
     }
   }
 }
